// src/app/models/masters/bank.ts
export class Bank {
    id: number;
    name: string;
    bankTypeId: number;
    countryId: number;
    bankType: string;
    country: string;
    accountCode: string;
    swiftCode: string;
    sudebanCode: string;
    active: number;
    binaryImage: string;
    image: string;
    createdByUserId: number;
}